Add files via upload
I cleaned and organized the two biggest data sets that we are planning on using. I included all of the variables that I was unsure about and figured I could remove them later in R if needed. I still had a couple questions about variables in the data sets so we should be good.